(INGEST-0): Fix typescript issues
